refactor(ws): refactor asyn
